pythononwheels/src/requirements.txt
alembic
appdirs
backports-abc
Cerberus
certifi
chardet
Click
colorama
dicttoxml
dnspython
#Flask
#Flask-Compress
idna
itsdangerous
jsonschema
junit2html
Mako
MarkupSafe
marshmallow
marshmallow-sqlalchemy
packaging
plotly
py
pymongo
PyMySQL
pyparsing
pytest
pytest-html
pytest-metadata
pytest-ordering
python-dateutil
python-editor
pytz
PyYAML
redis
requests
retrying
simplejson
six
SQLAlchemy==1.4.10
tinydb
tinydb-serialization
tornado
urllib3
Werkzeug==2.1.2
wheel
xmltodict
